+++ sme/src/main/java/sme/tech/innovators/sme/dto/response/PaymentInitDto.java
@@ -3,6 +3,8 @@ package sme.tech.innovators.sme.dto.response;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class PaymentInitDto {
@@ -11,6 +13,6 @@ public class PaymentInitDto {
     private String reference;
     private String publicKey;
     private String orderId;
-    private Integer amount;
+    private BigDecimal amount;
     private String currency;
 }
